Add new emitter and -Stream parameter
Adds the new emitter logic to reduce the amount of processing done on
each input object to YamlDotNet. This also paves the way for supporting
comments or other complex scenarios not normally exposed through the
Yaml* objects used before.

The ConvertTo-Yaml -Stream parameter can be used to serialize each input
object as a YAML string as they are received rather than collecting
until the end of the cmdlet call. This is useful if the cmdlet sending
the input is long running and the caller wishes to output the objects
immediately.